requirements.txt
selenium>=4.6.0
google-generativeai